Move component IPC calls to library
