+++ src/renderer/components/DraggableFile.tsx
@@ -1,6 +1,7 @@
 import styled from "@emotion/styled";
-import { ipcRenderer } from "electron";
 import React from "react";
+
+import { useHandleDragStart } from "@/lib/hooks/useHandleDragStart";
 
 const Outer = styled.div`
   color: inherit;
@@ -21,18 +22,13 @@ export interface DraggableFileProps {
  * such as copied to a different folder or dragged into web-sites etc.
  */
 export const DraggableFile: React.FC<DraggableFileProps> = ({ children, filePaths, className, style }) => {
-  const handleDragStart = (e: React.DragEvent<HTMLDivElement>) => {
-    e.preventDefault();
-    if (filePaths.length > 0) {
-      ipcRenderer.send("onDragStart", filePaths);
-    }
-  };
+  const handleDragStart = useHandleDragStart();
 
   return (
     <Outer
       className={className}
       style={style}
-      onDragStart={(e) => handleDragStart(e)}
+      onDragStart={(e) => handleDragStart(e, filePaths)}
       onClick={(e) => e.preventDefault()}
     >
       {children}
